pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "globule"
version = "1.0.0"
description = "Turn your scattered thoughts into structured drafts. Effortlessly."
authors = [{name = "Globule Team", email = "team@globule.dev"}]
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
keywords = ["ai", "notes", "knowledge-management", "semantic-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "click>=8.0.0",
    "asyncclick>=8.1.0",
    "textual>=0.40.0",
    "aiosqlite>=0.19.0",
    "aiohttp>=3.8.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "sqlite-vec>=0.0.2",
    "jsonschema>=4.17.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
clustering = [
    "hdbscan>=0.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "coverage>=7.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
globule = "globule.interfaces.cli.main:main"

[project.urls]
"Homepage" = "https://github.com/asavschaeffer/globule"
"Bug Reports" = "https://github.com/asavschaeffer/globule/issues"
"Source" = "https://github.com/asavschaeffer/globule"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
asyncio_mode = "auto"
addopts = "-v --tb=short"